refactor(agent-toolkit): modularize GraphQL operations for form tools
- Refactored the `FormQuestionsEditorTool` and `UpdateFormTool` classes to delegate GraphQL operations to dedicated helper classes (`FormQuestionsEditorToolHelpers` and `UpdateFormToolHelpers`).
- This change enhances code organization and maintainability by separating concerns and reducing complexity within the main tool classes.
- Action handlers for form question and update operations have been streamlined, improving readability and facilitating future enhancements.
